Clock refactoring
