+++ org.csstudio.display.builder.model/src/org/csstudio/display/builder/model/widgets/DigitalClockWidget.java
@@ -10,10 +10,8 @@ package org.csstudio.display.builder.model.widgets;
 
 
 import static org.csstudio.display.builder.model.properties.CommonWidgetProperties.newBooleanPropertyDescriptor;
-import static org.csstudio.display.builder.model.properties.CommonWidgetProperties.newStringPropertyDescriptor;
 
 import java.util.List;
-import java.util.Locale;
 
 import org.csstudio.display.builder.model.Messages;
 import org.csstudio.display.builder.model.Widget;
@@ -32,7 +30,7 @@ import org.csstudio.display.builder.model.properties.EnumWidgetProperty;
  * @version 1.0.0 23 Jan 2017
  */
 @SuppressWarnings("nls")
-public class DigitalClockWidget extends VisibleWidget {
+public class DigitalClockWidget extends BasicClockWidget {
 
     public static final WidgetDescriptor WIDGET_DESCRIPTOR = new WidgetDescriptor(
         "digital_clock",
@@ -86,40 +84,20 @@ public class DigitalClockWidget extends VisibleWidget {
         YOCTOPUCE
     }
 
-    public static final WidgetPropertyDescriptor<Design>  propLcdDesign         = new WidgetPropertyDescriptor<Design>(WidgetPropertyCategory.WIDGET,   "lcd_design",          Messages.WidgetProperties_LcdDesign) {
+    public static final WidgetPropertyDescriptor<Design>  propLcdDesign         = new WidgetPropertyDescriptor<Design>(WidgetPropertyCategory.WIDGET, "lcd_design",          Messages.WidgetProperties_LcdDesign) {
         @Override
         public EnumWidgetProperty<Design> createProperty ( Widget widget, Design defaultValue ) {
             return new EnumWidgetProperty<>(this, widget, defaultValue);
         }
     };
 
-    public static final WidgetPropertyDescriptor<Boolean> propDateVisible       = newBooleanPropertyDescriptor        (WidgetPropertyCategory.DISPLAY,  "date_visible",        Messages.WidgetProperties_DateVisible);
-    public static final WidgetPropertyDescriptor<Boolean> propSecondVisible     = newBooleanPropertyDescriptor        (WidgetPropertyCategory.DISPLAY,  "second_visible",      Messages.WidgetProperties_SecondVisible);
-    public static final WidgetPropertyDescriptor<String>  propTitle             = newStringPropertyDescriptor         (WidgetPropertyCategory.DISPLAY,  "title",               Messages.WidgetProperties_Title);
-    public static final WidgetPropertyDescriptor<Boolean> propTitleVisible      = newBooleanPropertyDescriptor        (WidgetPropertyCategory.DISPLAY,  "title_visible",       Messages.WidgetProperties_TitleVisible);
+    public static final WidgetPropertyDescriptor<Boolean> propLcdCrystalEnabled = newBooleanPropertyDescriptor        (WidgetPropertyCategory.MISC,   "lcd_crystal_enabled", Messages.WidgetProperties_LcdCrystalEnabled);
 
-    public static final WidgetPropertyDescriptor<Boolean> propRunning           = newBooleanPropertyDescriptor        (WidgetPropertyCategory.BEHAVIOR, "running",             Messages.WidgetProperties_Running);
-
-    public static final WidgetPropertyDescriptor<Boolean> propLcdCrystalEnabled = newBooleanPropertyDescriptor        (WidgetPropertyCategory.MISC,     "lcd_crystal_enabled", Messages.WidgetProperties_LcdCrystalEnabled);
-    public static final WidgetPropertyDescriptor<String>  propLocale            = newStringPropertyDescriptor         (WidgetPropertyCategory.MISC,     "locale",              Messages.WidgetProperties_Locale);
-    public static final WidgetPropertyDescriptor<Boolean> propShadowsEnabled    = newBooleanPropertyDescriptor        (WidgetPropertyCategory.MISC,     "shadows_enabled",     Messages.WidgetProperties_ShadowsEnabled);
-
-    private volatile WidgetProperty<Boolean> dateVisible;
     private volatile WidgetProperty<Boolean> lcdCrystalEnabled;
     private volatile WidgetProperty<Design>  lcdDesign;
-    private volatile WidgetProperty<String>  locale;
-    private volatile WidgetProperty<Boolean> running;
-    private volatile WidgetProperty<Boolean> secondVisible;
-    private volatile WidgetProperty<Boolean> shadowsEnabled;
-    private volatile WidgetProperty<String>  title;
-    private volatile WidgetProperty<Boolean> titleVisible;
 
     public DigitalClockWidget ( ) {
         super(WIDGET_DESCRIPTOR.getType(), 170, 90);
-    }
-
-    public WidgetProperty<Boolean> propDateVisible ( ) {
-        return dateVisible;
     }
 
     public WidgetProperty<Boolean> propLcdCrystalEnabled ( ) {
@@ -130,30 +108,6 @@ public class DigitalClockWidget extends VisibleWidget {
         return lcdDesign;
     }
 
-    public WidgetProperty<String> propLocale ( ) {
-        return locale;
-    }
-
-    public WidgetProperty<Boolean> propRunning ( ) {
-        return running;
-    }
-
-    public WidgetProperty<Boolean> propSecondVisible ( ) {
-        return secondVisible;
-    }
-
-    public WidgetProperty<Boolean> propShadowsEnabled ( ) {
-        return shadowsEnabled;
-    }
-
-    public WidgetProperty<String> propTitle ( ) {
-        return title;
-    }
-
-    public WidgetProperty<Boolean> propTitleVisible ( ) {
-        return titleVisible;
-    }
-
     @Override
     protected void defineProperties ( final List<WidgetProperty<?>> properties ) {
 
@@ -161,17 +115,7 @@ public class DigitalClockWidget extends VisibleWidget {
 
         properties.add(lcdDesign         = propLcdDesign.createProperty(this, Design.SECTIONS));
 
-        properties.add(dateVisible       = propDateVisible.createProperty(this, true));
-        properties.add(secondVisible     = propSecondVisible.createProperty(this, true));
-        properties.add(title             = propTitle.createProperty(this, ""));
-        properties.add(titleVisible      = propTitleVisible.createProperty(this, false));
-
-        properties.add(lcdCrystalEnabled = propLcdCrystalEnabled.createProperty(this, true));
-        properties.add(locale            = propLocale.createProperty(this, Locale.getDefault().toLanguageTag()));
-        properties.add(shadowsEnabled    = propShadowsEnabled.createProperty(this, true));
-
-        //  Properties not visible in the property sheet.
-        running = propRunning.createProperty(this, true);
+        properties.add(lcdCrystalEnabled = propLcdCrystalEnabled.createProperty(this, false));
 
     }
