+++ org.csstudio.display.builder.representation.javafx/src/org/csstudio/display/builder/representation/javafx/widgets/ClockRepresentation.java
@@ -9,11 +9,7 @@
 package org.csstudio.display.builder.representation.javafx.widgets;
 
 
-import static org.csstudio.display.builder.representation.ToolkitRepresentation.logger;
-
-import java.util.Locale;
 import java.util.Objects;
-import java.util.logging.Level;
 
 import org.csstudio.display.builder.model.DirtyFlag;
 import org.csstudio.display.builder.model.WidgetProperty;
@@ -22,7 +18,6 @@ import org.csstudio.display.builder.representation.javafx.JFXUtil;
 
 import eu.hansolo.medusa.Clock;
 import eu.hansolo.medusa.Clock.ClockSkinType;
-import eu.hansolo.medusa.ClockBuilder;
 import javafx.scene.paint.Color;
 import javafx.scene.paint.Paint;
 
@@ -32,14 +27,15 @@ import javafx.scene.paint.Paint;
  * @version 1.0.0 18 Jan 2017
  */
 @SuppressWarnings("nls")
-public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWidget> {
+public class ClockRepresentation extends BasicClockRepresentation<ClockWidget> {
 
     private final DirtyFlag dirtyBehavior = new DirtyFlag();
-    private final DirtyFlag dirtyGeometry = new DirtyFlag();
-    private final DirtyFlag dirtyLook = new DirtyFlag();
+    private final DirtyFlag dirtyLook     = new DirtyFlag();
 
     @Override
     public void updateChanges ( ) {
+
+        super.updateChanges();
 
         Object value;
 
@@ -62,27 +58,6 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
             if ( !Objects.equals(value, jfx_node.isDiscreteSeconds()) ) {
                 jfx_node.setDiscreteSeconds((boolean) value);
             }
-
-            value = model_widget.propRunning().getValue();
-
-            if ( !Objects.equals(value, jfx_node.isRunning()) ) {
-                jfx_node.setRunning((boolean) value);
-            }
-
-        }
-
-        if ( dirtyGeometry.checkAndClear() ) {
-
-            value = model_widget.propVisible().getValue();
-
-            if ( !Objects.equals(value, jfx_node.isVisible()) ) {
-                jfx_node.setVisible((boolean) value);
-            }
-
-            jfx_node.setLayoutX(model_widget.propX().getValue());
-            jfx_node.setLayoutY(model_widget.propY().getValue());
-            jfx_node.setPrefWidth(model_widget.propWidth().getValue());
-            jfx_node.setPrefHeight(model_widget.propHeight().getValue());
 
         }
 
@@ -118,12 +93,6 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
                 jfx_node.setDateColor((Color) value);
             }
 
-            value = model_widget.propDateVisible().getValue();
-
-            if ( !Objects.equals(value, jfx_node.isDateVisible()) ) {
-                jfx_node.setDateVisible((boolean) value);
-            }
-
             value = JFXUtil.convert(model_widget.propHourColor().getValue());
 
             if ( !Objects.equals(value, jfx_node.getHourColor()) ) {
@@ -146,24 +115,6 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
 
             if ( !Objects.equals(value, jfx_node.getKnobColor()) ) {
                 jfx_node.setKnobColor((Color) value);
-            }
-
-            value = model_widget.propLocale().getValue();
-
-            if ( value != null ) {
-
-                Locale l = Locale.getDefault();
-
-                try {
-                    l = Locale.forLanguageTag(value.toString());
-                } catch ( Exception ex ) {
-                    logger.log(Level.WARNING, "Unable to convert \"{0}\" to a Local instance [{1}].", new Object[] { value.toString(), ex.getMessage()});
-                }
-
-                if ( !l.equals(jfx_node.getLocale()) ) {
-                    jfx_node.setLocale(l);
-                }
-
             }
 
             value = JFXUtil.convert(model_widget.propMinuteColor().getValue());
@@ -190,18 +141,6 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
                 jfx_node.setSecondColor((Color) value);
             }
 
-            value = model_widget.propSecondVisible().getValue();
-
-            if ( !Objects.equals(value, jfx_node.isSecondsVisible()) ) {
-                jfx_node.setSecondsVisible((boolean) value);
-            }
-
-            value = model_widget.propShadowsEnabled().getValue();
-
-            if ( !Objects.equals(value, jfx_node.getShadowsEnabled()) ) {
-                jfx_node.setShadowsEnabled((boolean) value);
-            }
-
             value = JFXUtil.convert(model_widget.propTextColor().getValue());
 
             if ( !Objects.equals(value, jfx_node.getTextColor()) ) {
@@ -226,22 +165,10 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
                 jfx_node.setTickLabelsVisible((boolean) value);
             }
 
-            value = model_widget.propTitle().getValue();
-
-            if ( !Objects.equals(value, jfx_node.getTitle()) ) {
-                jfx_node.setTitle((String) value);
-            }
-
             value = JFXUtil.convert(model_widget.propTitleColor().getValue());
 
             if ( !Objects.equals(value, jfx_node.getTitleColor()) ) {
                 jfx_node.setTitleColor((Color) value);
-            }
-
-            value = model_widget.propTitleVisible().getValue();
-
-            if ( !Objects.equals(value, jfx_node.isTitleVisible()) ) {
-                jfx_node.setTitleVisible((boolean) value);
             }
 
         }
@@ -251,50 +178,29 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
     @Override
     protected Clock createJFXNode ( ) throws Exception {
 
-        Clock clock = ClockBuilder.create()
-                                  .skinType(ClockSkinType.valueOf(model_widget.propSkin().getValue().name()))
-                                  .prefHeight(model_widget.propHeight().getValue())
-                                  .prefWidth(model_widget.propWidth().getValue())
-                                  //--------------------------------------------------------
-                                  //  Previous properties must be set first.
-                                  //--------------------------------------------------------
-                                  .backgroundPaint(model_widget.propTransparent().getValue() ? Color.TRANSPARENT : JFXUtil.convert(model_widget.propBackgroundColor().getValue()))
-                                  .borderPaint(JFXUtil.convert(model_widget.propBorderColor().getValue()))
-                                  .borderWidth(model_widget.propBorderWidth().getValue())
-                                  .dateColor(JFXUtil.convert(model_widget.propDateColor().getValue()))
-                                  .dateVisible(model_widget.propDateVisible().getValue())
-                                  .discreteHours(model_widget.propDiscreteHours().getValue())
-                                  .discreteMinutes(model_widget.propDiscreteMinutes().getValue())
-                                  .discreteSeconds(model_widget.propDiscreteSeconds().getValue())
-                                  .hourColor(JFXUtil.convert(model_widget.propHourColor().getValue()))
-                                  .hourTickMarkColor(JFXUtil.convert(model_widget.propHourTickMarkColor().getValue()))
-                                  .hourTickMarksVisible(model_widget.propHourTickMarkVisible().getValue())
-                                  .knobColor(JFXUtil.convert(model_widget.propKnobColor().getValue()))
-                                  .minuteColor(JFXUtil.convert(model_widget.propMinuteColor().getValue()))
-                                  .minuteTickMarkColor(JFXUtil.convert(model_widget.propMinuteTickMarkColor().getValue()))
-                                  .minuteTickMarksVisible(model_widget.propMinuteTickMarkVisible().getValue())
-                                  .running(model_widget.propRunning().getValue())
-                                  .secondColor(JFXUtil.convert(model_widget.propSecondColor().getValue()))
-                                  .secondsVisible(model_widget.propSecondVisible().getValue())
-                                  .shadowsEnabled(model_widget.propShadowsEnabled().getValue())
-                                  .textColor(JFXUtil.convert(model_widget.propTextColor().getValue()))
-                                  .textVisible(model_widget.propTextVisible().getValue())
-                                  .tickLabelColor(JFXUtil.convert(model_widget.propTickLabelColor().getValue()))
-                                  .tickLabelsVisible(model_widget.propTickLabelVisible().getValue())
-                                  .title(model_widget.propTitle().getValue())
-                                  .titleColor(JFXUtil.convert(model_widget.propTitleColor().getValue()))
-                                  .titleVisible(model_widget.propTitleVisible().getValue())
-                                  .build();
+        Clock clock = super.createJFXNode();
 
-        String locale = model_widget.propLocale().getValue();
-
-        if ( locale != null ) {
-            try {
-                jfx_node.setLocale(Locale.forLanguageTag(locale.toString()));
-            } catch ( Exception ex ) {
-                logger.log(Level.WARNING, "Unable to convert \"{0}\" to a Local instance [{1}].", new Object[] { locale.toString(), ex.getMessage()});
-            }
-        }
+        clock.setSkinType(ClockSkinType.valueOf(model_widget.propSkin().getValue().name()));
+        clock.setBackgroundPaint(model_widget.propTransparent().getValue() ? Color.TRANSPARENT : JFXUtil.convert(model_widget.propBackgroundColor().getValue()));
+        clock.setBorderPaint(JFXUtil.convert(model_widget.propBorderColor().getValue()));
+        clock.setBorderWidth(model_widget.propBorderWidth().getValue());
+        clock.setDateColor(JFXUtil.convert(model_widget.propDateColor().getValue()));
+        clock.setDiscreteHours(model_widget.propDiscreteHours().getValue());
+        clock.setDiscreteMinutes(model_widget.propDiscreteMinutes().getValue());
+        clock.setDiscreteSeconds(model_widget.propDiscreteSeconds().getValue());
+        clock.setHourColor(JFXUtil.convert(model_widget.propHourColor().getValue()));
+        clock.setHourTickMarkColor(JFXUtil.convert(model_widget.propHourTickMarkColor().getValue()));
+        clock.setHourTickMarksVisible(model_widget.propHourTickMarkVisible().getValue());
+        clock.setKnobColor(JFXUtil.convert(model_widget.propKnobColor().getValue()));
+        clock.setMinuteColor(JFXUtil.convert(model_widget.propMinuteColor().getValue()));
+        clock.setMinuteTickMarkColor(JFXUtil.convert(model_widget.propMinuteTickMarkColor().getValue()));
+        clock.setMinuteTickMarksVisible(model_widget.propMinuteTickMarkVisible().getValue());
+        clock.setSecondColor(JFXUtil.convert(model_widget.propSecondColor().getValue()));
+        clock.setTextColor(JFXUtil.convert(model_widget.propTextColor().getValue()));
+        clock.setTextVisible(model_widget.propTextVisible().getValue());
+        clock.setTickLabelColor(JFXUtil.convert(model_widget.propTickLabelColor().getValue()));
+        clock.setTickLabelsVisible(model_widget.propTickLabelVisible().getValue());
+        clock.setTitleColor(JFXUtil.convert(model_widget.propTitleColor().getValue()));
 
         clock.backgroundPaintProperty().addListener( ( s, o, n ) -> {
             if ( !Objects.equals(n, JFXUtil.convert(model_widget.propBackgroundColor().getValue())) ) {
@@ -314,11 +220,6 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
         clock.dateColorProperty().addListener( ( s, o, n ) -> {
             if ( !Objects.equals(n, JFXUtil.convert(model_widget.propDateColor().getValue())) ) {
                 model_widget.propDateColor().setValue(JFXUtil.convert(n));
-            }
-        });
-        clock.dateVisibleProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propDateVisible().getValue()) ) {
-                model_widget.propDateVisible().setValue(n);
             }
         });
         clock.discreteHoursProperty().addListener( ( s, o, n ) -> {
@@ -356,21 +257,6 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
                 model_widget.propKnobColor().setValue(JFXUtil.convert(n));
             }
         });
-        clock.localeProperty().addListener( ( s, o, n ) -> {
-            if ( n != null && !Objects.equals(n.toLanguageTag(), model_widget.propLocale().getValue()) ) {
-                model_widget.propLocale().setValue(n.toLanguageTag());
-            }
-        });
-        clock.layoutXProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propX().getValue()) ) {
-                model_widget.propX().setValue(n.intValue());
-            }
-        });
-        clock.layoutYProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propY().getValue()) ) {
-                model_widget.propY().setValue(n.intValue());
-            }
-        });
         clock.minuteColorProperty().addListener( ( s, o, n ) -> {
             if ( !Objects.equals(n, JFXUtil.convert(model_widget.propMinuteColor().getValue())) ) {
                 model_widget.propMinuteColor().setValue(JFXUtil.convert(n));
@@ -386,34 +272,9 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
                 model_widget.propMinuteTickMarkVisible().setValue(n);
             }
         });
-        clock.prefHeightProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propHeight().getValue()) ) {
-                model_widget.propHeight().setValue(n.intValue());
-            }
-        });
-        clock.prefWidthProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propWidth().getValue()) ) {
-                model_widget.propWidth().setValue(n.intValue());
-            }
-        });
-        clock.runningProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propRunning().getValue()) ) {
-                model_widget.propRunning().setValue(n);
-            }
-        });
         clock.secondColorProperty().addListener( ( s, o, n ) -> {
             if ( !Objects.equals(n, JFXUtil.convert(model_widget.propSecondColor().getValue())) ) {
                 model_widget.propSecondColor().setValue(JFXUtil.convert(n));
-            }
-        });
-        clock.secondsVisibleProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propSecondVisible().getValue()) ) {
-                model_widget.propSecondVisible().setValue(n);
-            }
-        });
-        clock.shadowsEnabledProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propShadowsEnabled().getValue()) ) {
-                model_widget.propShadowsEnabled().setValue(n);
             }
         });
         clock.textColorProperty().addListener( ( s, o, n ) -> {
@@ -436,19 +297,9 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
                 model_widget.propTickLabelVisible().setValue(n);
             }
         });
-        clock.titleProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propTitle().getValue()) ) {
-                model_widget.propTitle().setValue(n);
-            }
-        });
         clock.titleColorProperty().addListener( ( s, o, n ) -> {
             if ( !Objects.equals(n, JFXUtil.convert(model_widget.propTitleColor().getValue())) ) {
                 model_widget.propTitleColor().setValue(JFXUtil.convert(n));
-            }
-        });
-        clock.titleVisibleProperty().addListener( ( s, o, n ) -> {
-            if ( !Objects.equals(n, model_widget.propTitleVisible().getValue()) ) {
-                model_widget.propTitleVisible().setValue(n);
             }
         });
 
@@ -459,52 +310,36 @@ public class ClockRepresentation extends RegionBaseRepresentation<Clock, ClockWi
     @Override
     protected void registerListeners ( ) {
 
-        model_widget.propVisible().addUntypedPropertyListener(this::geometryChanged);
-        model_widget.propX().addUntypedPropertyListener(this::geometryChanged);
-        model_widget.propY().addUntypedPropertyListener(this::geometryChanged);
-        model_widget.propWidth().addUntypedPropertyListener(this::geometryChanged);
-        model_widget.propHeight().addUntypedPropertyListener(this::geometryChanged);
+        super.registerListeners();
 
         model_widget.propBackgroundColor().addUntypedPropertyListener(this::lookChanged);
         model_widget.propBorderColor().addUntypedPropertyListener(this::lookChanged);
         model_widget.propBorderWidth().addUntypedPropertyListener(this::lookChanged);
         model_widget.propDateColor().addUntypedPropertyListener(this::lookChanged);
-        model_widget.propDateVisible().addUntypedPropertyListener(this::lookChanged);
         model_widget.propHourColor().addUntypedPropertyListener(this::lookChanged);
         model_widget.propHourTickMarkColor().addUntypedPropertyListener(this::lookChanged);
         model_widget.propHourTickMarkVisible().addUntypedPropertyListener(this::lookChanged);
         model_widget.propKnobColor().addUntypedPropertyListener(this::lookChanged);
-        model_widget.propLocale().addUntypedPropertyListener(this::lookChanged);
         model_widget.propMinuteColor().addUntypedPropertyListener(this::lookChanged);
         model_widget.propMinuteTickMarkColor().addUntypedPropertyListener(this::lookChanged);
         model_widget.propMinuteTickMarkVisible().addUntypedPropertyListener(this::lookChanged);
         model_widget.propSecondColor().addUntypedPropertyListener(this::lookChanged);
-        model_widget.propSecondVisible().addUntypedPropertyListener(this::lookChanged);
-        model_widget.propShadowsEnabled().addUntypedPropertyListener(this::lookChanged);
         model_widget.propSkin().addUntypedPropertyListener(this::lookChanged);
         model_widget.propTextColor().addUntypedPropertyListener(this::lookChanged);
         model_widget.propTextVisible().addUntypedPropertyListener(this::lookChanged);
         model_widget.propTickLabelColor().addUntypedPropertyListener(this::lookChanged);
         model_widget.propTickLabelVisible().addUntypedPropertyListener(this::lookChanged);
-        model_widget.propTitle().addUntypedPropertyListener(this::lookChanged);
         model_widget.propTitleColor().addUntypedPropertyListener(this::lookChanged);
-        model_widget.propTitleVisible().addUntypedPropertyListener(this::lookChanged);
         model_widget.propTransparent().addUntypedPropertyListener(this::lookChanged);
 
         model_widget.propDiscreteHours().addUntypedPropertyListener(this::behaviorChanged);
         model_widget.propDiscreteMinutes().addUntypedPropertyListener(this::behaviorChanged);
         model_widget.propDiscreteSeconds().addUntypedPropertyListener(this::behaviorChanged);
-        model_widget.propRunning().addUntypedPropertyListener(this::behaviorChanged);
 
     }
 
     private void behaviorChanged ( final WidgetProperty<?> property, final Object old_value, final Object new_value ) {
         dirtyBehavior.mark();
-        toolkit.scheduleUpdate(this);
-    }
-
-    private void geometryChanged ( final WidgetProperty<?> property, final Object old_value, final Object new_value ) {
-        dirtyGeometry.mark();
         toolkit.scheduleUpdate(this);
     }
